Fix mobile unsupported types: - block.Block - ChainConfig
